+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `withPlayerButtonOnImage` props and update style on
+  `SimpleCard` component.
+
 ## [1.19.0] - 2019-03-05
 
 Feature: